src/sim/requirements.txt
requests
pydispatcher
numpy
uvicorn
fastapi
matplotlib
pyorbital
pystac-client
odc-stac
click
cartopy
httplib2
pyqt6